+++ src/internal/superjson/mutate.ts
@@ -1,15 +1,16 @@
 import { err, ok, Result } from '../../lib';
-import { isEmptyRecord } from '../interpreter/variables';
 import {
-  normalizeProfileProviderDefaults,
-  normalizeProfileSettings,
-  normalizeUsecaseDefaults,
-} from './normalize';
+  castToNonPrimitive,
+  isEmptyRecord,
+  mergeVariables,
+} from '../interpreter/variables';
+import { normalizeProfileSettings } from './normalize';
 import {
   composeFileURI,
   isFileURIString,
   isVersionString,
   ProfileEntry,
+  ProfileProviderDefaults,
   ProfileProviderEntry,
   ProfileProviderSettings,
   ProfileSettings,
@@ -21,6 +22,68 @@ import {
   UsecaseDefaults,
 } from './schema';
 
+export function addProfileDefaults(
+  document: SuperJsonDocument,
+  profileName: string,
+  payload: UsecaseDefaults
+): boolean {
+  // if specified profile is not found
+  if (document.profiles === undefined) {
+    document.profiles = {};
+  }
+  if (document.profiles[profileName] === undefined) {
+    document.profiles[profileName] = '0.0.0';
+  }
+
+  const targetedProfile = document.profiles[profileName];
+
+  // if specified profile has shorthand notation
+  let defaults: UsecaseDefaults | undefined;
+  if (typeof targetedProfile === 'string') {
+    defaults = payload;
+
+    if (isVersionString(targetedProfile)) {
+      document.profiles[profileName] = {
+        version: targetedProfile,
+        defaults,
+      };
+
+      return true;
+    }
+
+    if (isFileURIString(targetedProfile)) {
+      document.profiles[profileName] = {
+        file: targetedProfile,
+        defaults,
+      };
+
+      return true;
+    }
+  } else {
+    if (targetedProfile.defaults) {
+      //Merge existing with new
+      defaults = mergeVariables(
+        castToNonPrimitive(targetedProfile.defaults) || {},
+        castToNonPrimitive(payload) || {}
+      ) as UsecaseDefaults;
+      document.profiles[profileName] = {
+        ...targetedProfile,
+        defaults,
+      };
+
+      return true;
+    } else {
+      document.profiles[profileName] = {
+        ...targetedProfile,
+        defaults: payload,
+      };
+
+      return true;
+    }
+  }
+
+  return false;
+}
 export function addProfile(
   document: SuperJsonDocument,
   profileName: string,
@@ -97,10 +160,15 @@ export function addProfile(
     defaults = payload.defaults;
   } else {
     if (targetedProfile.defaults) {
-      defaults = normalizeUsecaseDefaults(
-        payload.defaults,
-        normalizeUsecaseDefaults(targetedProfile.defaults)
-      );
+      if (!payload.defaults) {
+        defaults = targetedProfile.defaults;
+      } else {
+        //Merge existing with new
+        defaults = mergeVariables(
+          castToNonPrimitive(targetedProfile.defaults) || {},
+          castToNonPrimitive(payload.defaults) || {}
+        ) as UsecaseDefaults;
+      }
     }
     if (targetedProfile.priority) {
       priority = targetedProfile.priority;
@@ -126,7 +194,19 @@ export function addProfile(
 
   return true;
 }
+function resolvePriorityAddition(
+  existingPriority: string[] | undefined,
+  newProvider: string
+): string[] {
+  if (!existingPriority || existingPriority.length === 0) {
+    return [newProvider];
+  }
+  if (!existingPriority.includes(newProvider)) {
+    return [...existingPriority, newProvider];
+  }
 
+  return existingPriority;
+}
 export function addProfileProvider(
   document: SuperJsonDocument,
   profileName: string,
@@ -165,10 +245,10 @@ export function addProfileProvider(
       [providerName]: payload,
     };
 
-    targetedProfile.priority = [
-      ...(targetedProfile.priority || []),
-      providerName,
-    ];
+    targetedProfile.priority = resolvePriorityAddition(
+      targetedProfile.priority,
+      providerName
+    );
 
     return true;
   }
@@ -181,6 +261,10 @@ export function addProfileProvider(
       isEmptyRecord(profileProvider.defaults ?? {})
     ) {
       targetedProfile.providers[providerName] = composeFileURI(payload);
+      targetedProfile.priority = resolvePriorityAddition(
+        targetedProfile.priority,
+        providerName
+      );
 
       return true;
     }
@@ -189,19 +273,50 @@ export function addProfileProvider(
       file: trimFileURI(payload),
       defaults: profileProvider.defaults,
     };
+    targetedProfile.priority = resolvePriorityAddition(
+      targetedProfile.priority,
+      providerName
+    );
 
     return true;
   }
 
   // Priority #2: keep previous structure and merge
-  let defaults: UsecaseDefaults | undefined;
+  let defaults: ProfileProviderDefaults | undefined;
   if (typeof profileProvider === 'string') {
+    //Change
     defaults = payload.defaults;
-  } else if (profileProvider.defaults) {
-    defaults = normalizeProfileProviderDefaults(
-      payload.defaults,
-      normalizeUsecaseDefaults(targetedProfile.defaults)
+  } else {
+    if (profileProvider.defaults && payload.defaults) {
+      //Change
+      //Merge existing with new
+      defaults = mergeVariables(
+        castToNonPrimitive(profileProvider.defaults) || {},
+        castToNonPrimitive(payload.defaults) || {}
+      ) as ProfileProviderDefaults;
+    } else if (!profileProvider.defaults && payload.defaults) {
+      defaults = payload.defaults;
+    }
+  }
+  // if there no other keys and we changed defaults
+  if (
+    !('file' in payload) &&
+    !('mapVariant' in payload) &&
+    !('mapRevision' in payload) &&
+    defaults
+  ) {
+    targetedProfile.providers[providerName] = {
+      ...(typeof profileProvider === 'string'
+        ? { file: profileProvider }
+        : profileProvider),
+      defaults,
+    };
+    targetedProfile.priority = resolvePriorityAddition(
+      targetedProfile.priority,
+      providerName
     );
+
+    return true;
   }
 
   // when specified profile provider has file & defaults
@@ -210,6 +325,10 @@ export function addProfileProvider(
       ...payload,
       defaults,
     };
+    targetedProfile.priority = resolvePriorityAddition(
+      targetedProfile.priority,
+      providerName
+    );
 
     return true;
   }
@@ -221,6 +340,10 @@ export function addProfileProvider(
         ...payload,
         defaults,
       };
+      targetedProfile.priority = resolvePriorityAddition(
+        targetedProfile.priority,
+        providerName
+      );
 
       return true;
     }
@@ -240,6 +363,10 @@ export function addProfileProvider(
       ...mapProperties,
       defaults,
     };
+    targetedProfile.priority = resolvePriorityAddition(
+      targetedProfile.priority,
+      providerName
+    );
 
     return true;
   }