+++ src/internal/superjson/superjson.test.ts
@@ -22,6 +22,7 @@ import {
   ProviderEntry,
   SecurityValues,
   trimFileURI,
+  UsecaseDefaults,
 } from './schema';
 import { SuperJson } from './superjson';
 
@@ -232,6 +233,37 @@ describe('SuperJson', () => {
             "version": "1.0.Z",
             "providers": {
               "acme": {
+                "mapVariant": "my-bugfix",
+                "mapRevision": "1113"
+              }
+            }
+          }
+        },
+        "providers": {
+          "acme": {
+            "security": [
+              {
+                "id": "myApiKey",
+                "apikey": "SECRET"
+              }
+            ]
+          }
+        }
+      }`);
+      expect(SuperJson.loadSync('test').isErr()).toEqual(true);
+    });
+
+    it('returns err when there is an error during parsing super.json - defaults missing', () => {
+      mocked(statSync).mockReturnValue(mockStats);
+      mocked(readFileSync).mockReturnValue(`{
+        "profiles": {
+          "send-message": {
+            "version": "1.0.0",
+            "providers": {
+              "acme": {
+                "RetrieveCharacterInformation": {
+                  "retryPolicy": "circuit-breaker"
+                },
                 "mapVariant": "my-bugfix",
                 "mapRevision": "1113"
               }
@@ -691,11 +723,6 @@ describe('SuperJson', () => {
         },
         "providers": {
           "swapidev": {
-            "deployments": {
-              "default": {
-                "baseUrl": "https://www.some.differentUrl"
-              }
-            },
             "security": [
               {
                 "id": "myBasicAuth",
@@ -714,15 +741,6 @@ describe('SuperJson', () => {
           },
           "twillio": {
             "security": []
-          }
-        },
-        "lock": {
-          "starwars/character-information@^1.1": {
-            "version": "1.1.5",
-            "resolved": "https://store.superface.ai/profile/starwars/character-information@1.1.5",
-              "integrity": "sha512-0NKGC8Nf/4vvDpWKB7bwxIazvNnNHnZBX6XlyBXNl+fW8tpTef3PNMJMSErTz9LFnuv61vsKbc36u/Ek2YChWg==",
-            "astResolved": "",
-            "astIntegrity": "sha512-0NKGC8Nf/4vvDpWKB7bwxIazvNnNHnZBX6XlyBXNl+fW8tpTef3PNMJMSErTz9LFnuv61vsKbc36u/Ek2YChWg=="
           }
         }
       }`;
@@ -1161,6 +1179,208 @@ describe('SuperJson', () => {
             ],
           },
         },
+      });
+    });
+  });
+  describe('when adding profile defaults', () => {
+    it('adds profile deafults to empty super.json multiple times', () => {
+      const mockProfileName = 'profile';
+      const mockUseCaseName = 'usecase';
+
+      let mockProfileDeafultsEntry: UsecaseDefaults = {
+        [mockUseCaseName]: { providerFailover: false, input: { test: 'test' } },
+      };
+
+      expect(
+        superjson.addProfileDefaults(mockProfileName, mockProfileDeafultsEntry)
+      ).toEqual(true);
+      expect(superjson.document.profiles?.[mockProfileName]).toEqual({
+        defaults: {
+          [mockUseCaseName]: {
+            providerFailover: false,
+            input: { test: 'test' },
+          },
+        },
+        version: '0.0.0',
+      });
+
+      mockProfileDeafultsEntry = {
+        [mockUseCaseName]: {
+          providerFailover: true,
+          input: { test: 'new-test' },
+        },
+      };
+
+      expect(
+        superjson.addProfileDefaults(mockProfileName, mockProfileDeafultsEntry)
+      ).toEqual(true);
+      expect(superjson.document.profiles?.[mockProfileName]).toEqual({
+        defaults: {
+          [mockUseCaseName]: {
+            providerFailover: true,
+            input: { test: 'new-test' },
+          },
+        },
+        version: '0.0.0',
+      });
+    });
+
+    it('adds profile deafults to super.json with profile using uri path multiple times', () => {
+      const mockProfileName = 'profile';
+      const mockUseCaseName = 'usecase';
+
+      superjson = new SuperJson({
+        profiles: { [mockProfileName]: 'file://some/path' },
+      });
+
+      let mockProfileDeafultsEntry: UsecaseDefaults = {
+        [mockUseCaseName]: { providerFailover: false, input: { test: 'test' } },
+      };
+
+      expect(
+        superjson.addProfileDefaults(mockProfileName, mockProfileDeafultsEntry)
+      ).toEqual(true);
+      expect(superjson.document.profiles?.[mockProfileName]).toEqual({
+        defaults: {
+          [mockUseCaseName]: {
+            providerFailover: false,
+            input: { test: 'test' },
+          },
+        },
+        file: 'file://some/path',
+      });
+
+      mockProfileDeafultsEntry = {
+        [mockUseCaseName]: {
+          providerFailover: true,
+          input: { test: 'new-test' },
+        },
+      };
+
+      expect(
+        superjson.addProfileDefaults(mockProfileName, mockProfileDeafultsEntry)
+      ).toEqual(true);
+      expect(superjson.document.profiles?.[mockProfileName]).toEqual({
+        defaults: {
+          [mockUseCaseName]: {
+            providerFailover: true,
+            input: { test: 'new-test' },
+          },
+        },
+        file: 'file://some/path',
+      });
+    });
+
+    it('adds profile deafults to super.json with existing profile multiple times', () => {
+      const mockProfileName = 'profile';
+      const mockUseCaseName = 'usecase';
+
+      superjson = new SuperJson({
+        profiles: {
+          [mockProfileName]: {
+            version: '1.0.0',
+            priority: ['test'],
+            providers: { test: {} },
+          },
+        },
+      });
+
+      let mockProfileDeafultsEntry: UsecaseDefaults = {
+        [mockUseCaseName]: { providerFailover: true, input: { test: 'test' } },
+      };
+
+      expect(
+        superjson.addProfileDefaults(mockProfileName, mockProfileDeafultsEntry)
+      ).toEqual(true);
+      expect(superjson.document.profiles?.[mockProfileName]).toEqual({
+        defaults: {
+          [mockUseCaseName]: {
+            providerFailover: true,
+            input: { test: 'test' },
+          },
+        },
+        version: '1.0.0',
+        priority: ['test'],
+        providers: { test: {} },
+      });
+
+      mockProfileDeafultsEntry = {
+        [mockUseCaseName]: {
+          providerFailover: false,
+          input: { test: 'new-test' },
+        },
+      };
+
+      expect(
+        superjson.addProfileDefaults(mockProfileName, mockProfileDeafultsEntry)
+      ).toEqual(true);
+      expect(superjson.document.profiles?.[mockProfileName]).toEqual({
+        defaults: {
+          [mockUseCaseName]: {
+            providerFailover: false,
+            input: { test: 'new-test' },
+          },
+        },
+        version: '1.0.0',
+        priority: ['test'],
+        providers: { test: {} },
+      });
+    });
+
+    it('adds profile deafults to super.json with existing profile and existing defaults multiple times', () => {
+      const mockProfileName = 'profile';
+      const mockUseCaseName = 'usecase';
+
+      superjson = new SuperJson({
+        profiles: {
+          [mockProfileName]: {
+            version: '1.0.0',
+            priority: ['test'],
+            defaults: { [mockUseCaseName]: { providerFailover: false } },
+            providers: { test: {} },
+          },
+        },
+      });
+
+      let mockProfileDeafultsEntry: UsecaseDefaults = {
+        [mockUseCaseName]: { providerFailover: true, input: { test: 'test' } },
+      };
+
+      expect(
+        superjson.addProfileDefaults(mockProfileName, mockProfileDeafultsEntry)
+      ).toEqual(true);
+      expect(superjson.document.profiles?.[mockProfileName]).toEqual({
+        defaults: {
+          [mockUseCaseName]: {
+            providerFailover: true,
+            input: { test: 'test' },
+          },
+        },
+        version: '1.0.0',
+        priority: ['test'],
+        providers: { test: {} },
+      });
+
+      mockProfileDeafultsEntry = {
+        [mockUseCaseName]: {
+          providerFailover: false,
+          input: { test: 'new-test' },
+        },
+      };
+
+      expect(
+        superjson.addProfileDefaults(mockProfileName, mockProfileDeafultsEntry)
+      ).toEqual(true);
+      expect(superjson.document.profiles?.[mockProfileName]).toEqual({
+        defaults: {
+          [mockUseCaseName]: {
+            providerFailover: false,
+            input: { test: 'new-test' },
+          },
+        },
+        version: '1.0.0',
+        priority: ['test'],
+        providers: { test: {} },
       });
     });
   });
@@ -1931,7 +2151,7 @@ describe('SuperJson', () => {
       });
     });
 
-    it('returns false if super.json wasnt updated', () => {
+    it('adds profile provider to super.json with exisitng profile provider but without defaults', () => {
       const mockProfileName = 'profile';
       const mockProviderName = 'provider';
       const mockProfileProviderEntry: ProfileProviderEntry = {
@@ -1946,12 +2166,50 @@ describe('SuperJson', () => {
             providers: {
               [mockProviderName]: {
                 file: 'provider/path',
-                defaults: {
-                  input: {
-                    input: { test: 'test' },
-                    retryPolicy: { kind: OnFail.NONE },
-                  },
-                },
+              },
+            },
+          },
+        },
+      });
+
+      expect(
+        superjson.addProfileProvider(
+          mockProfileName,
+          mockProviderName,
+          mockProfileProviderEntry
+        )
+      ).toEqual(true);
+      expect(superjson.normalized.profiles[mockProfileName]).toEqual({
+        defaults: {},
+        file: 'some/path',
+        priority: [mockProviderName],
+        providers: {
+          [mockProviderName]: {
+            defaults: {
+              input: {
+                input: { test: 'test' },
+                retryPolicy: { kind: OnFail.NONE },
+              },
+            },
+            file: 'provider/path',
+          },
+        },
+      });
+    });
+
+    it('returns false if super.json wasnt updated', () => {
+      const mockProfileName = 'profile';
+      const mockProviderName = 'provider';
+      const mockProfileProviderEntry: ProfileProviderEntry = {};
+
+      superjson = new SuperJson({
+        profiles: {
+          [mockProfileName]: {
+            defaults: {},
+            file: 'some/path',
+            providers: {
+              [mockProviderName]: {
+                file: 'provider/path',
               },
             },
           },
@@ -1971,12 +2229,7 @@ describe('SuperJson', () => {
         priority: [mockProviderName],
         providers: {
           [mockProviderName]: {
-            defaults: {
-              input: {
-                input: { test: 'test' },
-                retryPolicy: { kind: OnFail.NONE },
-              },
-            },
+            defaults: {},
             file: 'provider/path',
           },
         },